Refactored some of the namings
