+++ API/__init__.py
@@ -28,6 +28,8 @@ def get_app() -> Flask:
         Message().register_all()
         from matches import MatchBP
         MatchBP().register_all()
+        from swipe import SwipeBP
+        SwipeBP().register_all()
 
     @app.teardown_appcontext
     def rm_sess(exception=None):
@@ -38,6 +40,5 @@ def get_app() -> Flask:
 
 if __name__ == '__main__':
     app = get_app()
-    print(app.url_map)
     CORS(app, supports_credentials=True, origins="*")
     app.run(debug=app.config['DEBUG'], host="0.0.0.0", port=3001)
