+++ API/db.py
@@ -57,7 +57,8 @@ def load_orms(app: Flask) -> None:
     from auth import User
     from profile_imp import Profile
     from message import MessageTable
-    from matches import MatchTable, SwipeTable
+    from matches import MatchTable
+    from swipe import SwipeTable
     db_inf: DBComps = app.config['DB']
     db_inf['base'].metadata.create_all(bind=db_inf['engine'])
